refactor(catalog): isolate cache lifecycle
